Refactor: Merge document and document-history resolvers and gql typedefs
